requirements.txt
aiofauna==0.0.11
  - aiohttp [required: >=3.8.4,<4.0.0, installed: 3.8.4]
    - aiosignal [required: >=1.1.2, installed: 1.3.1]
      - frozenlist [required: >=1.1.0, installed: 1.3.3]
    - async-timeout [required: >=4.0.0a3,<5.0, installed: 4.0.2]
    - attrs [required: >=17.3.0, installed: 22.2.0]
    - charset-normalizer [required: >=2.0,<4.0, installed: 3.1.0]
    - frozenlist [required: >=1.1.1, installed: 1.3.3]
    - multidict [required: >=4.5,<7.0, installed: 6.0.4]
    - yarl [required: >=1.0,<2.0, installed: 1.8.2]
      - idna [required: >=2.0, installed: 3.4]
      - multidict [required: >=4.0, installed: 6.0.4]
  - aiohttp-sse [required: >=2.1.0,<3.0.0, installed: 2.1.0]
    - aiohttp [required: >=3.0, installed: 3.8.4]
      - aiosignal [required: >=1.1.2, installed: 1.3.1]
        - frozenlist [required: >=1.1.0, installed: 1.3.3]
      - async-timeout [required: >=4.0.0a3,<5.0, installed: 4.0.2]
      - attrs [required: >=17.3.0, installed: 22.2.0]
      - charset-normalizer [required: >=2.0,<4.0, installed: 3.1.0]
      - frozenlist [required: >=1.1.1, installed: 1.3.3]
      - multidict [required: >=4.5,<7.0, installed: 6.0.4]
      - yarl [required: >=1.0,<2.0, installed: 1.8.2]
        - idna [required: >=2.0, installed: 3.4]
        - multidict [required: >=4.0, installed: 6.0.4]
  - iso8601 [required: >=1.1.0,<2.0.0, installed: 1.1.0]
  - Jinja2 [required: >=3.1.2,<4.0.0, installed: 3.1.2]
    - MarkupSafe [required: >=2.0, installed: 2.1.2]
  - pydantic [required: >=1.10.7,<2.0.0, installed: 1.10.7]
    - typing-extensions [required: >=4.2.0, installed: 4.5.0]
email-validator==2.0.0
  - dnspython [required: >=2.0.0, installed: 2.3.0]
  - idna [required: >=2.0.0, installed: 3.4]
python-dotenv==1.0.0
